requirements.txt
aiogram>=3.20
aiohttp
aiohttp-socks
aiosqlite
python-dotenv
